backend/requirements.txt
fastapi>=0.100.0
pydantic-settings>=2.0.0
uvicorn>=0.23.0
python-multipart>=0.0.6
supabase>=1.0.3
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
